@@ -19,8 +19,8 @@ plugins {
   id 'net.ltgt.errorprone'
 }
 
-sourceCompatibility = JavaVersion.VERSION_1_7
-targetCompatibility = JavaVersion.VERSION_1_7
+sourceCompatibility = JavaVersion.VERSION_1_8
+targetCompatibility = JavaVersion.VERSION_1_8
 
 dependencies {
   annotationProcessor deps.build.nullAway
