@@ -17,9 +17,7 @@
 package com.uber.autodispose.lifecycle;
 
 import com.uber.autodispose.AutoDisposePlugins;
-import com.uber.autodispose.OutsideScopeException;
 import io.reactivex.CompletableSource;
-import io.reactivex.functions.Consumer;
 import io.reactivex.observers.TestObserver;
 import io.reactivex.subjects.PublishSubject;
 import java.util.Comparator;
@@ -41,7 +39,8 @@ public final class LifecycleScopesTest {
 
     try {
       testSource(resolveScopeFromLifecycle(lifecycle));
-      throw new AssertionError("Lifecycle resolution should have failed due to missing start event");
+      throw new AssertionError("Lifecycle resolution should have failed due to missing start "
+          + "event");
     } catch (LifecycleNotStartedException ignored) {
 
     }
@@ -72,10 +71,8 @@ public final class LifecycleScopesTest {
   @Test public void lifecycleCheckEnd_shouldFailIfEndedWithHandler() {
     TestLifecycleScopeProvider lifecycle = TestLifecycleScopeProvider.createInitial(STOPPED);
 
-    AutoDisposePlugins.setOutsideScopeHandler(new Consumer<OutsideScopeException>() {
-      @Override public void accept(OutsideScopeException e) {
-        // Swallow the exception.
-      }
+    AutoDisposePlugins.setOutsideScopeHandler(e -> {
+      // Swallow the exception.
     });
 
     testSource(resolveScopeFromLifecycle(lifecycle, true)).assertComplete();
@@ -85,11 +82,9 @@ public final class LifecycleScopesTest {
     TestLifecycleScopeProvider lifecycle = TestLifecycleScopeProvider.createInitial(STOPPED);
 
     final RuntimeException expected = new RuntimeException("Expected");
-    AutoDisposePlugins.setOutsideScopeHandler(new Consumer<OutsideScopeException>() {
-      @Override public void accept(OutsideScopeException e) {
-        // Throw it back
-        throw expected;
-      }
+    AutoDisposePlugins.setOutsideScopeHandler(e -> {
+      // Throw it back
+      throw expected;
     });
 
     testSource(resolveScopeFromLifecycle(lifecycle, true)).assertError(expected);
@@ -98,7 +93,8 @@ public final class LifecycleScopesTest {
   @Test public void lifecycleCheckEndDisabled_shouldRouteThroughOnError() {
     TestLifecycleScopeProvider lifecycle = TestLifecycleScopeProvider.createInitial(STOPPED);
 
-    testSource(resolveScopeFromLifecycle(lifecycle, false)).assertError(LifecycleEndedException.class);
+    testSource(resolveScopeFromLifecycle(lifecycle,
+        false)).assertError(LifecycleEndedException.class);
   }
 
   @Test public void resolveScopeFromLifecycle_normal() {
@@ -179,7 +175,8 @@ public final class LifecycleScopesTest {
   @Test public void resolveScopeFromLifecycle_normal_comparable() {
     PublishSubject<NegativeComparableInteger> lifecycle = PublishSubject.create();
 
-    TestObserver<?> o = testSource(resolveScopeFromLifecycle(lifecycle, new NegativeComparableInteger(3)));
+    TestObserver<?> o =
+        testSource(resolveScopeFromLifecycle(lifecycle, new NegativeComparableInteger(3)));
 
     lifecycle.onNext(new NegativeComparableInteger(-1));
     o.assertNotTerminated();
@@ -196,11 +193,7 @@ public final class LifecycleScopesTest {
   @Test public void resolveScopeFromLifecycle_normal_comparator() {
     PublishSubject<Integer> lifecycle = PublishSubject.create();
 
-    Comparator<Integer> comparator = new Comparator<Integer>() {
-      @Override public int compare(Integer o1, Integer o2) {
-        return Integer.compare(-o1, o2);
-      }
-    };
+    Comparator<Integer> comparator = (o1, o2) -> Integer.compare(-o1, o2);
 
     TestObserver<?> o = testSource(resolveScopeFromLifecycle(lifecycle, 3, comparator));
 
