@@ -20,13 +20,8 @@ import com.uber.autodispose.observers.AutoDisposingMaybeObserver;
 import com.uber.autodispose.test.RecordingObserver;
 import com.uber.autodispose.test.RxErrorsRule;
 import io.reactivex.Maybe;
-import io.reactivex.MaybeEmitter;
 import io.reactivex.MaybeObserver;
-import io.reactivex.MaybeOnSubscribe;
-import io.reactivex.functions.BiFunction;
-import io.reactivex.functions.Cancellable;
 import io.reactivex.functions.Consumer;
-import io.reactivex.functions.Predicate;
 import io.reactivex.observers.TestObserver;
 import io.reactivex.plugins.RxJavaPlugins;
 import io.reactivex.subjects.CompletableSubject;
@@ -37,16 +32,16 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import static com.google.common.truth.Truth.assertThat;
+import static com.uber.autodispose.AutoDispose.autoDisposable;
 import static com.uber.autodispose.TestUtil.makeProvider;
 import static com.uber.autodispose.TestUtil.outsideScopeProvider;
 
 public class AutoDisposeMaybeObserverTest {
 
-  private static final RecordingObserver.Logger LOGGER = new RecordingObserver.Logger() {
-    @Override public void log(String message) {
-      System.out.println(AutoDisposeMaybeObserverTest.class.getSimpleName() + ": " + message);
-    }
-  };
+  private static final RecordingObserver.Logger LOGGER =
+      message -> System.out.println(AutoDisposeMaybeObserverTest.class.getSimpleName()
+          + ": "
+          + message);
 
   @Rule public RxErrorsRule rule = new RxErrorsRule();
 
@@ -54,7 +49,7 @@ public class AutoDisposeMaybeObserverTest {
     RecordingObserver<Integer> o = new RecordingObserver<>(LOGGER);
     MaybeSubject<Integer> source = MaybeSubject.create();
     CompletableSubject scope = CompletableSubject.create();
-    source.as(AutoDispose.<Integer>autoDisposable(scope))
+    source.as(autoDisposable(scope))
         .subscribe(o);
     o.takeSubscribe();
 
@@ -73,11 +68,9 @@ public class AutoDisposeMaybeObserverTest {
 
   @Test public void autoDispose_withSuperClassGenerics_compilesFine() {
     Maybe.just(new BClass())
-        .as(AutoDispose.<BClass>autoDisposable(ScopeProvider.UNBOUND))
-        .subscribe(new Consumer<AClass>() {
-          @Override public void accept(AClass aClass) {
+        .as(autoDisposable(ScopeProvider.UNBOUND))
+        .subscribe((Consumer<AClass>) aClass -> {
 
-          }
         });
   }
 
@@ -85,13 +78,11 @@ public class AutoDisposeMaybeObserverTest {
     RecordingObserver<Integer> o = new RecordingObserver<>(LOGGER);
     MaybeSubject<Integer> source = MaybeSubject.create();
     CompletableSubject scope = CompletableSubject.create();
-    source.as(AutoDispose.<Integer>autoDisposable(scope))
+    source.as(autoDisposable(scope))
         .subscribe(o);
-    source.as(AutoDispose.<Integer>autoDisposable(scope))
-        .subscribe(new Consumer<Integer>() {
-          @Override public void accept(Integer integer) {
+    source.as(autoDisposable(scope))
+        .subscribe(integer -> {
 
-          }
         });
     o.takeSubscribe();
 
@@ -113,7 +104,7 @@ public class AutoDisposeMaybeObserverTest {
     MaybeSubject<Integer> source = MaybeSubject.create();
     CompletableSubject scope = CompletableSubject.create();
     ScopeProvider provider = makeProvider(scope);
-    source.as(AutoDispose.<Integer>autoDisposable(provider))
+    source.as(autoDisposable(provider))
         .subscribe(o);
     o.takeSubscribe();
 
@@ -133,7 +124,7 @@ public class AutoDisposeMaybeObserverTest {
     MaybeSubject<Integer> source = MaybeSubject.create();
     CompletableSubject scope = CompletableSubject.create();
     ScopeProvider provider = makeProvider(scope);
-    source.as(AutoDispose.<Integer>autoDisposable(provider))
+    source.as(autoDisposable(provider))
         .subscribe(o);
     o.takeSubscribe();
 
@@ -153,7 +144,7 @@ public class AutoDisposeMaybeObserverTest {
     MaybeSubject<Integer> source = MaybeSubject.create();
     CompletableSubject scope = CompletableSubject.create();
     ScopeProvider provider = makeProvider(scope);
-    source.as(AutoDispose.<Integer>autoDisposable(provider))
+    source.as(autoDisposable(provider))
         .subscribe(o);
     o.takeSubscribe();
 
@@ -175,19 +166,17 @@ public class AutoDisposeMaybeObserverTest {
     final AtomicReference<MaybeObserver> atomicObserver = new AtomicReference<>();
     final AtomicReference<MaybeObserver> atomicAutoDisposingObserver = new AtomicReference<>();
     try {
-      RxJavaPlugins.setOnMaybeSubscribe(new BiFunction<Maybe, MaybeObserver, MaybeObserver>() {
-        @Override public MaybeObserver apply(Maybe source, MaybeObserver observer) {
-          if (atomicObserver.get() == null) {
-            atomicObserver.set(observer);
-          } else if (atomicAutoDisposingObserver.get() == null) {
-            atomicAutoDisposingObserver.set(observer);
-            RxJavaPlugins.setOnObservableSubscribe(null);
-          }
-          return observer;
+      RxJavaPlugins.setOnMaybeSubscribe((source, observer) -> {
+        if (atomicObserver.get() == null) {
+          atomicObserver.set(observer);
+        } else if (atomicAutoDisposingObserver.get() == null) {
+          atomicAutoDisposingObserver.set(observer);
+          RxJavaPlugins.setOnObservableSubscribe(null);
         }
+        return observer;
       });
       Maybe.just(1)
-          .as(AutoDispose.<Integer>autoDisposable(ScopeProvider.UNBOUND))
+          .as(autoDisposable(ScopeProvider.UNBOUND))
           .subscribe();
 
       assertThat(atomicAutoDisposingObserver.get()).isNotNull();
@@ -203,17 +192,9 @@ public class AutoDisposeMaybeObserverTest {
   @Test public void verifyCancellation() {
     final AtomicInteger i = new AtomicInteger();
     //noinspection unchecked because Java
-    Maybe<Integer> source = Maybe.create(new MaybeOnSubscribe<Integer>() {
-      @Override public void subscribe(MaybeEmitter<Integer> e) {
-        e.setCancellable(new Cancellable() {
-          @Override public void cancel() {
-            i.incrementAndGet();
-          }
-        });
-      }
-    });
+    Maybe<Integer> source = Maybe.create(e -> e.setCancellable(i::incrementAndGet));
     CompletableSubject scope = CompletableSubject.create();
-    source.as(AutoDispose.<Integer>autoDisposable(scope))
+    source.as(autoDisposable(scope))
         .subscribe();
 
     assertThat(i.get()).isEqualTo(0);
@@ -228,7 +209,7 @@ public class AutoDisposeMaybeObserverTest {
 
   @Test public void autoDispose_withScopeProviderCompleted_shouldNotReportDoubleSubscriptions() {
     TestObserver<Object> o = MaybeSubject.create()
-        .as(AutoDispose.autoDisposable(ScopeProvider.UNBOUND))
+        .as(autoDisposable(ScopeProvider.UNBOUND))
         .test();
     o.assertNoValues();
     o.assertNoErrors();
@@ -238,7 +219,7 @@ public class AutoDisposeMaybeObserverTest {
 
   @Test public void unbound_shouldStillPassValues() {
     MaybeSubject<Integer> s = MaybeSubject.create();
-    TestObserver<Integer> o = s.as(AutoDispose.<Integer>autoDisposable(ScopeProvider.UNBOUND))
+    TestObserver<Integer> o = s.as(autoDisposable(ScopeProvider.UNBOUND))
         .test();
 
     s.onSuccess(1);
@@ -246,12 +227,10 @@ public class AutoDisposeMaybeObserverTest {
   }
 
   @Test public void autoDispose_outsideScope_withProviderAndNoOpPlugin_shouldFailSilently() {
-    AutoDisposePlugins.setOutsideScopeHandler(new Consumer<OutsideScopeException>() {
-      @Override public void accept(OutsideScopeException e) { }
-    });
+    AutoDisposePlugins.setOutsideScopeHandler(e -> { });
     ScopeProvider provider = outsideScopeProvider();
     MaybeSubject<Integer> source = MaybeSubject.create();
-    TestObserver<Integer> o = source.as(AutoDispose.<Integer>autoDisposable(provider))
+    TestObserver<Integer> o = source.as(autoDisposable(provider))
         .test();
 
     assertThat(source.hasObservers()).isFalse();
@@ -260,22 +239,17 @@ public class AutoDisposeMaybeObserverTest {
   }
 
   @Test public void autoDispose_outsideScope_withProviderAndPlugin_shouldFailWithWrappedExp() {
-    AutoDisposePlugins.setOutsideScopeHandler(new Consumer<OutsideScopeException>() {
-      @Override public void accept(OutsideScopeException e) {
-        // Wrap in an IllegalStateException so we can verify this is the exception we see on the
-        // other side
-        throw new IllegalStateException(e);
-      }
+    AutoDisposePlugins.setOutsideScopeHandler(e -> {
+      // Wrap in an IllegalStateException so we can verify this is the exception we see on the
+      // other side
+      throw new IllegalStateException(e);
     });
     ScopeProvider provider = outsideScopeProvider();
-    TestObserver<Integer> o = MaybeSubject.<Integer>create().as(AutoDispose.<Integer>autoDisposable(provider))
+    TestObserver<Integer> o = MaybeSubject.<Integer>create().as(autoDisposable(provider))
         .test();
 
     o.assertNoValues();
-    o.assertError(new Predicate<Throwable>() {
-      @Override public boolean test(Throwable throwable) {
-        return throwable instanceof IllegalStateException && throwable.getCause() instanceof OutsideScopeException;
-      }
-    });
+    o.assertError(throwable -> throwable instanceof IllegalStateException
+        && throwable.getCause() instanceof OutsideScopeException);
   }
 }
