@@ -21,12 +21,8 @@ import com.uber.autodispose.test.RxErrorsRule;
 import io.reactivex.BackpressureStrategy;
 import io.reactivex.Flowable;
 import io.reactivex.FlowableEmitter;
-import io.reactivex.FlowableOnSubscribe;
 import io.reactivex.disposables.Disposable;
-import io.reactivex.functions.BiFunction;
-import io.reactivex.functions.Cancellable;
 import io.reactivex.functions.Consumer;
-import io.reactivex.functions.Predicate;
 import io.reactivex.plugins.RxJavaPlugins;
 import io.reactivex.processors.PublishProcessor;
 import io.reactivex.subjects.CompletableSubject;
@@ -38,6 +34,7 @@ import org.junit.Test;
 import org.reactivestreams.Subscriber;
 
 import static com.google.common.truth.Truth.assertThat;
+import static com.uber.autodispose.AutoDispose.autoDisposable;
 import static com.uber.autodispose.TestUtil.outsideScopeProvider;
 
 public class AutoDisposeSubscriberTest {
@@ -48,7 +45,7 @@ public class AutoDisposeSubscriberTest {
     TestSubscriber<Integer> o = new TestSubscriber<>();
     PublishProcessor<Integer> source = PublishProcessor.create();
     CompletableSubject scope = CompletableSubject.create();
-    Disposable d = source.as(AutoDispose.<Integer>autoDisposable(scope))
+    Disposable d = source.as(autoDisposable(scope))
         .subscribeWith(o);
     o.assertSubscribed();
 
@@ -69,18 +66,16 @@ public class AutoDisposeSubscriberTest {
 
   @Test public void autoDispose_withSuperClassGenerics_compilesFine() {
     Flowable.just(new BClass())
-        .as(AutoDispose.<BClass>autoDisposable(ScopeProvider.UNBOUND))
-        .subscribe(new Consumer<AClass>() {
-          @Override public void accept(AClass aClass) {
+        .as(autoDisposable(ScopeProvider.UNBOUND))
+        .subscribe((Consumer<AClass>) aClass -> {
 
-          }
         });
   }
 
   @Test public void autoDispose_withMaybe_interrupted() {
     PublishProcessor<Integer> source = PublishProcessor.create();
     CompletableSubject scope = CompletableSubject.create();
-    TestSubscriber<Integer> o = source.as(AutoDispose.<Integer>autoDisposable(scope))
+    TestSubscriber<Integer> o = source.as(autoDisposable(scope))
         .test();
     o.assertSubscribed();
 
@@ -105,7 +100,7 @@ public class AutoDisposeSubscriberTest {
     PublishProcessor<Integer> source = PublishProcessor.create();
     CompletableSubject scope = CompletableSubject.create();
     ScopeProvider provider = TestUtil.makeProvider(scope);
-    TestSubscriber<Integer> o = source.as(AutoDispose.<Integer>autoDisposable(provider))
+    TestSubscriber<Integer> o = source.as(autoDisposable(provider))
         .test();
     o.assertSubscribed();
 
@@ -136,23 +131,21 @@ public class AutoDisposeSubscriberTest {
     final AtomicReference<Subscriber> atomicSubscriber = new AtomicReference<>();
     final AtomicReference<Subscriber> atomicAutoDisposingSubscriber = new AtomicReference<>();
     try {
-      RxJavaPlugins.setOnFlowableSubscribe(new BiFunction<Flowable, Subscriber, Subscriber>() {
-        @Override public Subscriber apply(Flowable source, Subscriber subscriber) {
-          if (atomicSubscriber.get() == null) {
-            System.out.println(subscriber.getClass()
-                .toString());
-            atomicSubscriber.set(subscriber);
-          } else if (atomicAutoDisposingSubscriber.get() == null) {
-            System.out.println(subscriber.getClass()
-                .toString());
-            atomicAutoDisposingSubscriber.set(subscriber);
-            RxJavaPlugins.setOnFlowableSubscribe(null);
-          }
-          return subscriber;
+      RxJavaPlugins.setOnFlowableSubscribe((source, subscriber) -> {
+        if (atomicSubscriber.get() == null) {
+          System.out.println(subscriber.getClass()
+              .toString());
+          atomicSubscriber.set(subscriber);
+        } else if (atomicAutoDisposingSubscriber.get() == null) {
+          System.out.println(subscriber.getClass()
+              .toString());
+          atomicAutoDisposingSubscriber.set(subscriber);
+          RxJavaPlugins.setOnFlowableSubscribe(null);
         }
+        return subscriber;
       });
       Flowable.just(1)
-          .as(AutoDispose.<Integer>autoDisposable(ScopeProvider.UNBOUND))
+          .as(autoDisposable(ScopeProvider.UNBOUND))
           .subscribe();
 
       assertThat(atomicAutoDisposingSubscriber.get()).isNotNull();
@@ -169,18 +162,12 @@ public class AutoDisposeSubscriberTest {
     final AtomicInteger i = new AtomicInteger();
     //noinspection unchecked because Java
     final FlowableEmitter<Integer>[] emitter = new FlowableEmitter[1];
-    Flowable<Integer> source = Flowable.create(new FlowableOnSubscribe<Integer>() {
-      @Override public void subscribe(FlowableEmitter<Integer> e) {
-        e.setCancellable(new Cancellable() {
-          @Override public void cancel() {
-            i.incrementAndGet();
-          }
-        });
-        emitter[0] = e;
-      }
+    Flowable<Integer> source = Flowable.create(e -> {
+      e.setCancellable(i::incrementAndGet);
+      emitter[0] = e;
     }, BackpressureStrategy.LATEST);
     CompletableSubject scope = CompletableSubject.create();
-    source.as(AutoDispose.<Integer>autoDisposable(scope))
+    source.as(autoDisposable(scope))
         .subscribe();
 
     assertThat(i.get()).isEqualTo(0);
@@ -198,7 +185,7 @@ public class AutoDisposeSubscriberTest {
 
   @Test public void autoDispose_withScopeProviderCompleted_shouldNotReportDoubleSubscriptions() {
     TestSubscriber<Object> o = PublishProcessor.create()
-        .as(AutoDispose.autoDisposable(ScopeProvider.UNBOUND))
+        .as(autoDisposable(ScopeProvider.UNBOUND))
         .test();
     o.assertNoValues();
     o.assertNoErrors();
@@ -208,7 +195,7 @@ public class AutoDisposeSubscriberTest {
 
   @Test public void unbound_shouldStillPassValues() {
     PublishProcessor<Integer> s = PublishProcessor.create();
-    TestSubscriber<Integer> o = s.as(AutoDispose.<Integer>autoDisposable(ScopeProvider.UNBOUND))
+    TestSubscriber<Integer> o = s.as(autoDisposable(ScopeProvider.UNBOUND))
         .test();
 
     s.onNext(1);
@@ -217,12 +204,10 @@ public class AutoDisposeSubscriberTest {
   }
 
   @Test public void autoDispose_outsideScope_withProviderAndNoOpPlugin_shouldFailSilently() {
-    AutoDisposePlugins.setOutsideScopeHandler(new Consumer<OutsideScopeException>() {
-      @Override public void accept(OutsideScopeException e) { }
-    });
+    AutoDisposePlugins.setOutsideScopeHandler(e -> { });
     ScopeProvider provider = outsideScopeProvider();
     PublishProcessor<Integer> source = PublishProcessor.create();
-    TestSubscriber<Integer> o = source.as(AutoDispose.<Integer>autoDisposable(provider))
+    TestSubscriber<Integer> o = source.as(autoDisposable(provider))
         .test();
 
     assertThat(source.hasSubscribers()).isFalse();
@@ -231,22 +216,17 @@ public class AutoDisposeSubscriberTest {
   }
 
   @Test public void autoDispose_outsideScope_withProviderAndPlugin_shouldFailWithWrappedExp() {
-    AutoDisposePlugins.setOutsideScopeHandler(new Consumer<OutsideScopeException>() {
-      @Override public void accept(OutsideScopeException e) {
-        // Wrap in an IllegalStateException so we can verify this is the exception we see on the
-        // other side
-        throw new IllegalStateException(e);
-      }
+    AutoDisposePlugins.setOutsideScopeHandler(e -> {
+      // Wrap in an IllegalStateException so we can verify this is the exception we see on the
+      // other side
+      throw new IllegalStateException(e);
     });
     ScopeProvider provider = outsideScopeProvider();
-    TestSubscriber<Integer> o = PublishProcessor.<Integer>create().as(AutoDispose.<Integer>autoDisposable(provider))
+    TestSubscriber<Integer> o = PublishProcessor.<Integer>create().as(autoDisposable(provider))
         .test();
 
     o.assertNoValues();
-    o.assertError(new Predicate<Throwable>() {
-      @Override public boolean test(Throwable throwable) {
-        return throwable instanceof IllegalStateException && throwable.getCause() instanceof OutsideScopeException;
-      }
-    });
+    o.assertError(throwable -> throwable instanceof IllegalStateException
+        && throwable.getCause() instanceof OutsideScopeException);
   }
 }
